+++ Products/BeakerSessionDataManager/sessiondata.py
@@ -205,5 +205,16 @@ class BeakerSessionDataObject(Implicit):
 
     _p_changed = property(_get_p_changed, _set_p_changed)
 
+    # dict API compatibility used by some plone addons e.g. collective.z3cform.wizard
+
+    def has_key(self, key):
+        return key in self.session.keys()
+
+    def get(self, key):
+        return self.session.get(key)
+
+    def __getitem__(self, key):
+        return self.get(key)
+
 
 InitializeClass(BeakerSessionDataObject)